Extract `Sample` to a file.
