+++ src/test_data.rs
@@ -75,37 +75,11 @@ impl TransferTarget {
     }
 }
 
-pub(crate) struct Sample<V> {
-    label: String,
-    sample: V,
-}
-
-impl<V> Sample<V> {
-    pub(crate) fn new(label: String, sample: V) -> Sample<V> {
-        Sample { label, sample }
-    }
-
-    pub(crate) fn destructure(self) -> (String, V) {
-        (self.label, self.sample)
-    }
-
-    pub(crate) fn map_sample<VV, F: FnOnce(V) -> VV>(self, f: F) -> Sample<VV> {
-        Sample {
-            label: self.label,
-            sample: f(self.sample),
-        }
-    }
-
-    pub(crate) fn add_label(&mut self, label: String) {
-        self.label = format!("{}-{}", self.label, label);
-    }
-}
-
 pub(crate) mod native_transfer {
     use casper_execution_engine::core::engine_state::ExecutableDeployItem;
     use casper_types::{AccessRights, URef, U512};
 
-    use crate::test_data::{Sample, TransferTarget};
+    use crate::{sample::Sample, test_data::TransferTarget};
 
     use super::NativeTransfer;
